split benchmarks into 3 actions
