+++ benchmarks/clickhouse/e2e.kafka.benchmark.ts
@@ -1,16 +1,16 @@
-import { LogLevel, PluginsServerConfig, Queue } from '../src/types'
-import { resetTestDatabase } from '../tests/helpers/sql'
-import { startPluginsServer } from '../src/server'
-import { makePiscina } from '../src/worker/piscina'
-import { PluginsServer } from '../src/types'
-import { createPosthog, DummyPostHog } from '../src/extensions/posthog'
-import { pluginConfig39 } from '../tests/helpers/plugins'
-import { delay, UUIDT } from '../src/utils'
-import { delayUntilEventIngested } from '../tests/shared/process-event'
+import { LogLevel, PluginsServerConfig, Queue } from '../../src/types'
+import { resetTestDatabase } from '../../tests/helpers/sql'
+import { startPluginsServer } from '../../src/server'
+import { makePiscina } from '../../src/worker/piscina'
+import { PluginsServer } from '../../src/types'
+import { createPosthog, DummyPostHog } from '../../src/extensions/posthog'
+import { pluginConfig39 } from '../../tests/helpers/plugins'
+import { delay, UUIDT } from '../../src/utils'
+import { delayUntilEventIngested } from '../../tests/shared/process-event'
 import { performance } from 'perf_hooks'
-import { resetKafka } from '../tests/helpers/kafka'
-import { KAFKA_EVENTS_PLUGIN_INGESTION } from '../src/ingestion/topics'
-import { resetTestDatabaseClickhouse } from '../tests/helpers/clickhouse'
+import { resetKafka } from '../../tests/helpers/kafka'
+import { KAFKA_EVENTS_PLUGIN_INGESTION } from '../../src/ingestion/topics'
+import { resetTestDatabaseClickhouse } from '../../tests/helpers/clickhouse'
 
 jest.setTimeout(600000) // 10min timeout
