+++ benchmarks/postgres/e2e.celery.benchmark.ts
@@ -1,12 +1,12 @@
-import { LogLevel, PluginsServerConfig, Queue } from '../src/types'
-import { resetTestDatabase } from '../tests/helpers/sql'
-import { startPluginsServer } from '../src/server'
-import { makePiscina } from '../src/worker/piscina'
-import { PluginsServer } from '../src/types'
-import { createPosthog, DummyPostHog } from '../src/extensions/posthog'
-import { pluginConfig39 } from '../tests/helpers/plugins'
-import { delay, UUIDT } from '../src/utils'
-import { delayUntilEventIngested } from '../tests/shared/process-event'
+import { LogLevel, PluginsServerConfig, Queue } from '../../src/types'
+import { resetTestDatabase } from '../../tests/helpers/sql'
+import { startPluginsServer } from '../../src/server'
+import { makePiscina } from '../../src/worker/piscina'
+import { PluginsServer } from '../../src/types'
+import { createPosthog, DummyPostHog } from '../../src/extensions/posthog'
+import { pluginConfig39 } from '../../tests/helpers/plugins'
+import { delay, UUIDT } from '../../src/utils'
+import { delayUntilEventIngested } from '../../tests/shared/process-event'
 import { performance } from 'perf_hooks'
 
 jest.setTimeout(600000) // 10min timeout
