+++ benchmarks/vm/worker.benchmark.ts
@@ -1,12 +1,12 @@
-import { makePiscina } from '../src/worker/piscina'
-import { defaultConfig } from '../src/config'
+import { makePiscina } from '../../src/worker/piscina'
+import { defaultConfig } from '../../src/config'
 import { PluginEvent } from '@posthog/plugin-scaffold/src/types'
 import { performance } from 'perf_hooks'
 import * as os from 'os'
-import { LogLevel } from '../src/types'
-import { resetTestDatabase } from '../tests/helpers/sql'
+import { LogLevel } from '../../src/types'
+import { resetTestDatabase } from '../../tests/helpers/sql'
 
-jest.mock('../src/sql')
+jest.mock('../../src/sql')
 jest.setTimeout(600000) // 600 sec timeout
 
 function processOneEvent(
@@ -110,7 +110,7 @@ test('piscina worker benchmark', async () => {
                 async function processEvent (event, meta) {
                     await new Promise(resolve => __jestSetTimeout(() => resolve(), 100))
                     event.properties = { "somewhere": "over the rainbow" };
-                    return event             
+                    return event
                 }
             `,
         },
