+++ package.json
@@ -8,7 +8,10 @@
         "test": "jest --runInBand tests/**/*.test.ts",
         "test:postgres": "jest --runInBand tests/postgres/*.test.ts tests/*.test.ts",
         "test:clickhouse": "jest --runInBand tests/clickhouse/*.test.ts",
-        "benchmark": "node --expose-gc node_modules/.bin/jest --runInBand benchmarks/",
+        "benchmark": "yarn run benchmarks:clickhouse && yarn run benchmark:postgres && yarn run benchmarks:vm",
+        "benchmark:clickhouse": "node --expose-gc node_modules/.bin/jest --runInBand benchmarks/vm/",
+        "benchmark:postgres": "node --expose-gc node_modules/.bin/jest --runInBand benchmarks/clickhouse/",
+        "benchmark:vm": "node --expose-gc node_modules/.bin/jest --runInBand benchmarks/postgres/",
         "start": "yarn start:dev",
         "start:dist": "node dist/src/index.js --base-dir ../posthog",
         "start:dev": "ts-node-dev --exit-child src/index.ts --base-dir ../posthog",
